+++ learn/views.py
@@ -1,5 +1,0 @@
-from django.views.generic import TemplateView
-
-
-class LearnPageView(TemplateView):
-    template_name = 'learn/learn_page.html'