+++ react-components/src/Services.js
@@ -1,5 +1,9 @@
+/* eslint-disable */
 const MESSAGE_UNEXPECTED_ERROR = {'__all__': ['Unexpected Error']}
 const MESSAGE_PERMISSION_DENIED = {'__all__': ['You do not have permission to perform this action']}
+const MESSAGE_NOT_FOUND_ERROR = {'__all__': ['Not found']}
+const MESSAGE_TIMEOUT_ERROR = {'__all__': ['Request timed out']}
+const MESSAGE_BAD_REQUEST_ERROR = {'__all__': ['Bad request']}
 
 
 const post = function(url, data) {
@@ -13,6 +17,29 @@ const post = function(url, data) {
     },
     body: JSON.stringify(data),
   })
+}
+
+const get = function(url, params) {
+  const parsedUrl = new URL(`${location.origin}${url}`)
+  const parsedParams = new URLSearchParams(params).toString();
+  parsedUrl.search = parsedParams
+
+  return fetch(parsedUrl, {
+    method: 'get',
+    headers: {
+      'Accept': 'application/json',
+      'X-CSRFToken': config.csrfToken,
+      'X-Requested-With': 'XMLHttpRequest',
+    },
+  })
+}
+
+const getCountryData = function(country) {
+  return get(config.countryDataUrl, {'country': country}).then(response => responseHandler(response).json())
+}
+
+const lookupProduct = function({q}) {
+  return get(config.apiLookupProductUrl, {q}).then(response => responseHandler(response).json())
 }
 
 
@@ -31,8 +58,8 @@ const checkCredentials = function({ email, password }) {
 }
 
 
-const updateCompany = function({ company_name, expertise_industries, expertise_countries, first_name, last_name }) {
-  const data = { company_name, expertise_industries, expertise_countries, first_name, last_name }
+const updateCompany = function({ expertise_industries, expertise_countries, expertise_products_services }) {
+  const data = { expertise_industries, expertise_countries, expertise_products_services }
   return post(config.apiUpdateCompanyUrl, data).then(responseHandler)
 }
 
@@ -41,16 +68,26 @@ const responseHandler = function(response) {
     return response.json().then(error => { throw error })
   } else if (response.status == 403) {
     throw MESSAGE_PERMISSION_DENIED
+  } else if (response.status == 404) {
+    throw MESSAGE_NOT_FOUND_ERROR
+  } else if (response.status == 504) {
+    throw MESSAGE_TIMEOUT_ERROR
+  } else if (response.status == 400) {
+    throw MESSAGE_BAD_REQUEST_ERROR
   } else if (response.status != 200) {
     throw MESSAGE_UNEXPECTED_ERROR
+  } else {
+    return response
   }
 }
 
 // static values that will not change during execution of the code
 let config = {}
 const setConfig = function({
+  countryDataUrl,
   apiLoginUrl,
   apiSignupUrl,
+  apiLookupProductUrl,
   apiUpdateCompanyUrl,
   countryOptions,
   csrfToken,
@@ -64,9 +101,12 @@ const setConfig = function({
   userCountries,
   userIndustries,
   verifyCodeUrl,
+  userIsAuthenticated,
 }) {
+  config.countryDataUrl = countryDataUrl
   config.apiLoginUrl = apiLoginUrl
   config.apiSignupUrl = apiSignupUrl
+  config.apiLookupProductUrl = apiLookupProductUrl
   config.apiUpdateCompanyUrl = apiUpdateCompanyUrl
   config.countryOptions = countryOptions
   config.csrfToken = csrfToken
@@ -80,6 +120,7 @@ const setConfig = function({
   config.verifyCodeUrl = verifyCodeUrl
   config.userCountries = userCountries
   config.userIndustries = userIndustries
+  config.userIsAuthenticated = userIsAuthenticated
 }
 
 export default {
@@ -87,10 +128,14 @@ export default {
   checkCredentials,
   checkVerificationCode,
   updateCompany,
+  getCountryData,
+  lookupProduct,
   setConfig,
   config,
   messages: {
     MESSAGE_UNEXPECTED_ERROR,
     MESSAGE_PERMISSION_DENIED,
+    MESSAGE_NOT_FOUND_ERROR,
+    MESSAGE_TIMEOUT_ERROR,
   }
 }