+++ react-components/src/components/Modal.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import React from 'react'
 import PropTypes from 'prop-types'
 import ReactModal from 'react-modal'
@@ -7,21 +8,30 @@ import { withCookies, useCookies } from 'react-cookie';
 
 export function Modal(props){
   const [cookies, setCookie] = useCookies([props.skipFeatureCookieName])
-  const SkipFeature = cookies[props.skipFeatureCookieName] == 'true'
-  const [isOpen, setIsOpen] = React.useState(!SkipFeature && props.isOpen)
+  
+  function isOpen() {
+    // some modals are opened on user click. Those should be able to skip the
+    // "do not open if user previosuly asked not to see the modal again"
+    if (props.performSkipFeatureCookieCheck) {
+      const skipFeature = cookies[props.skipFeatureCookieName] == 'true'
+      return !skipFeature && props.isOpen
+    }
+    return props.isOpen
+  }
+
 
   function handleClose(event){
     event.preventDefault()
-    setIsOpen(false)
+    props.setIsOpen(false)
   }
 
   function handleRequestSkipFeature() {
-    setCookie(props.skipFeatureCookieName, 'true');
-    setIsOpen(false)
+    setCookie(props.skipFeatureCookieName, 'true', {path: '/'})
+    props.setIsOpen(false)
   }
 
   function getSkipFeature() {
-    const SkipFeature= props.skipFeatureComponent
+    const SkipFeature = props.skipFeatureComponent
     if (SkipFeature) {
       return <SkipFeature onClick={handleRequestSkipFeature} />
     }
@@ -29,7 +39,7 @@ export function Modal(props){
 
   return (
     <ReactModal
-      isOpen={isOpen}
+      isOpen={isOpen()}
       onRequestClose={handleClose}
       className={'ReactModal__Content ReactModalCentreScreen ' + props.className}
       overlayClassName='ReactModal__Overlay ReactModalCentreScreen'
@@ -47,11 +57,13 @@ export function Modal(props){
 Modal.propTypes = {
   isOpen: PropTypes.bool,
   skipFeatureCookieName: PropTypes.string,
-  id: PropTypes.string
+  id: PropTypes.string,
+  performSkipFeatureCookieCheck: PropTypes.bool,
 }
 
 Modal.defaultProps = {
   isOpen: false,
+  performSkipFeatureCookieCheck: true,
 }
 
 export default withCookies(Modal)