+++ react-components/tests/views/CountriesModal/Base.test.jsx
@@ -6,18 +6,16 @@ import { mount, shallow } from 'enzyme'
 import Enzyme from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
 
-import {Base, STEP_SUCCESS, STEP_COUNTRIES} from '@src/views/CountriesModal/Base'
+import { Base, STEP_SUCCESS, STEP_COUNTRIES } from '@src/views/CountriesModal/Base'
 import StepSuccess from '@src/views/CountriesModal/StepSuccess'
 import StepCountries from '@src/views/CountriesModal/StepCountries'
 import Services from '@src/Services'
 
-
 Enzyme.configure({ adapter: new Adapter() })
 
-jest.mock('@src/Services');
+jest.mock('@src/Services')
 
 const createEvent = () => ({ preventDefault: jest.fn() })
-
 
 beforeEach(() => {
   jest.useFakeTimers()
@@ -38,7 +36,6 @@ const defaultProps = {
 }
 
 describe('Countries Base', () => {
-
   const { assign } = window.location
 
   beforeEach(() => {
@@ -50,9 +47,9 @@ describe('Countries Base', () => {
     window.location.assign = assign
   })
 
-  test('bad data results in errors passed down', done => {
+  test('bad data results in errors passed down', (done) => {
     // given the credentials are incorrect
-    const errors = {'expertise_countries': ['You are denied']}
+    const errors = { expertise_countries: ['You are denied'] }
     const countries = ['some country']
 
     Services.updateCompany.mockImplementation(() => Promise.reject(errors))
@@ -70,11 +67,11 @@ describe('Countries Base', () => {
     })
   })
 
-  test('end to end country select', done => {
+  test('end to end country select', (done) => {
     // given the credentials are correct
     Services.updateCompany.mockImplementation(() => Promise.resolve())
     // and the user is on the sectors step
-    const props = {...defaultProps, currentStep: STEP_COUNTRIES}
+    const props = { ...defaultProps, currentStep: STEP_COUNTRIES }
     const component = mount(<Base {...props} />)
 
     // when the user chooses "some industry"