+++ tests/unit/core/test_views.py
@@ -3,14 +3,13 @@ from unittest.mock import patch, Mock
 import json
 
 import pytest
+from directory_api_client import api_client
 
 from django.urls import reverse
 
-from core import helpers
+from core import helpers, models, serializers
 from tests.helpers import create_response
-from directory_api_client import api_client
 from tests.unit.learn.factories import LessonPageFactory, TopicPageFactory
-from learn.models import LessonViewHit
 
 
 @pytest.fixture
@@ -38,6 +37,22 @@ def test_api_update_company_success(mock_update_company_profile, mock_get_compan
         },
         sso_session_id=user.session_id,
     )
+
+
+@pytest.mark.django_db
+def test_api_update_company_too_many_companies(client, user):
+    company_data = {
+        'expertise_countries': json.dumps(['USA', 'China', 'Australia', 'New Zealand']),
+    }
+
+    client.force_login(user)
+
+    response = client.post(reverse('core:api-update-company'), company_data)
+
+    assert response.status_code == 400
+    assert response.json() == {
+        'expertise_countries': [serializers.CompanySerializer.MESSAGE_TOO_MANY_COUNTRIES],
+    }
 
 
 @pytest.mark.django_db
@@ -106,20 +121,20 @@ def test_dashboard_page_lesson_progress(
     client.force_login(user)
 
     # given the user has read some lessons
-    topic_one = TopicPageFactory(parent=domestic_homepage, slug='topic-one')
-    topic_two = TopicPageFactory(parent=domestic_homepage, slug='topic-two')
+    topic_one = TopicPageFactory(parent=domestic_homepage, slug='topic-one', record_read_progress=True)
+    topic_two = TopicPageFactory(parent=domestic_homepage, slug='topic-two', record_read_progress=True)
     lesson_one = LessonPageFactory(parent=topic_one, slug='lesson-one')
     lesson_two = LessonPageFactory(parent=topic_one, slug='lesson-two')
-    LessonPageFactory(parent=topic_one, slug='lesson-three')
+    LessonPageFactory(parent=topic_one, slug='lesson-three',)
     LessonPageFactory(parent=topic_one, slug='lesson-four')
-    LessonViewHit.objects.create(
-        lesson=lesson_one,
-        topic=topic_one,
+    models.PageView.objects.create(
+        page=lesson_one,
+        list_page=topic_one,
         sso_id=user.pk
     )
-    LessonViewHit.objects.create(
-        lesson=lesson_two,
-        topic=topic_one,
+    models.PageView.objects.create(
+        page=lesson_two,
+        list_page=topic_one,
         sso_id=user.pk
     )
 
@@ -129,13 +144,13 @@ def test_dashboard_page_lesson_progress(
 
     # then the progress is exposed
     assert response.status_code == 200
-    assert len(response.context_data['topics']) == 2
-    assert response.context_data['topics'][0] == topic_one
-    assert response.context_data['topics'][0].read_count == 2
-    assert response.context_data['topics'][0].read_progress == 50
-    assert response.context_data['topics'][1] == topic_two
-    assert response.context_data['topics'][1].read_count == 0
-    assert response.context_data['topics'][1].read_progress is None
+    assert len(response.context_data['list_pages']) == 2
+    assert response.context_data['list_pages'][0] == topic_one
+    assert response.context_data['list_pages'][0].read_count == 2
+    assert response.context_data['list_pages'][0].read_progress == 50
+    assert response.context_data['list_pages'][1] == topic_two
+    assert response.context_data['list_pages'][1].read_count == 0
+    assert response.context_data['list_pages'][1].read_progress is None
 
 
 @pytest.mark.django_db
@@ -305,3 +320,29 @@ def test_markets_not_logged_in(mock_get_company_profile, client):
     assert response.status_code == 200
     assert response.context_data['page_title'] is None
     assert response.context_data['most_popular_countries'] is None
+
+
+@mock.patch.object(helpers, 'search_commodity_by_term')
+def test_search_commodity_by_term(mock_search_commodity_by_term, client):
+    mock_search_commodity_by_term.return_value = data = [
+        {'value': '123323', 'label': 'some description'},
+        {'value': '223323', 'label': 'some other description'},
+    ]
+    term = 'some term'
+
+    response = client.get(reverse('core:api-lookup-product'), {'q': term})
+
+    assert response.status_code == 200
+    assert response.json() == data
+    assert mock_search_commodity_by_term.call_count == 1
+    assert mock_search_commodity_by_term.call_args == mock.call(term=term)
+
+
+@pytest.mark.django_db
+def test_list_page_uses_right_template(domestic_homepage, rf, user):
+    request = rf.get('/')
+    request.user = user
+    topic_page = TopicPageFactory(parent=domestic_homepage)
+    lesson_page = LessonPageFactory(parent=topic_page)
+    response = lesson_page.serve(request)
+    assert response.template_name == 'learn/lesson_page.html'