+++ CHANGELOG.md
@@ -53,4 +53,12 @@
 - MVP-356 - marketdata target-markets
 - MVP-357 - Implement new signup modal design
 - MVP-362 - styling and correct HTML for country statistics view
+- MVP-289 - add products modal to export plan
 - MVP-360 - muliple markets
+- MVP-382 - add product modal to lesson page
+- MVP-350 - Capabilities learn how to export carousel
+- MVP-385 - Consolidate page models
+- MVP-335 - Add market capture component
+- no ticket - rm redundant CircleCI step setup_remote_docker
+- MVP-388 - add commdity name
+